🎨 refactor(artwork): transform the artwork page method to a composable
- ♻️ update the method to get the authors to accept multiple author

- get the username of an author if he don't have nickname

- JSDOC and try catch implementation

- no test for the composable atm
